Add files via upload
Client had a problem where in the upload command sent to server, size of file and nonce was present. 

That meant the file must have been either compressed and encrypted or only encrypted in advanced, before even getting user's input settings for the upload. There was no way to know the size of the output file.

So we used to perform one of these actions (compress-encrypt or encrypt) first, sending the server all the information. Then we waited for user to enter the proper settings for the upload. Because user could have changed the name of the file on server, we needed the user to complete entering the desired settings and then communicate with the server to see if filename already exists.

Now client will send the upload command, without file size and nonce, wait for user to enter settings for the upload, then search the local database if the file already exists (to prevent communicating with the server) and at last compress-encrypt or encrypt, send file size and nonce and the file.

Actions 'compression-encryption' and 'encryption' happens in the background, allowing user to perform extra actions.